+++ web/package.json
@@ -3,7 +3,10 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@material-ui/core": "^3.9.3",
+        "@material-ui/icons": "^3.0.2",
         "bootstrap": "^4.3.1",
+        "material-ui": "^0.20.2",
         "react": "^16.8.6",
         "react-dom": "^16.8.6",
         "react-router-dom": "^5.0.0",
