+++ web/src/index.css
@@ -1,55 +1,3 @@
 body {
 	font-family: 'Roboto Condensed', sans-serif;
 }
-
-/* GLOBAL*/
-html,
-body {
-	font: 16px/1.5 Arial, Helvetica, sans-serif;
-	margin: 0;
-	padding: 0;
-	height: 100vh;
-	width: 100%;
-	/*background-color: #f4f4f4;*/
-}
-
-.container {
-	display: block;
-	width: 80%;
-	margin: 0 10%;
-	padding: 0;
-	overflow: hidden;
-}
-
-#sign_in_header {
-	width: 100%;
-	height: 70px;
-	background-color: red;
-}
-
-#logo_header {
-	border: solid 2px black;
-	display: inline-block;
-	/*float: left;*/
-	margin-top: 5px;
-	margin-bottom: 5px;
-}
-
-#login {
-	float: right;
-	line-height: 60px;
-	/*border:solid 1px black;*/
-}
-
-#username_login {
-	display: inline-block;
-	/*border:solid 1px black;*/
-}
-
-#password_login {
-	display: inline-block;
-	/*border:solid 1px black;*/
-}
-
-/* Media Rules*/
-@media only screen and(max-width:900px) {}
